minor refactor of managing default search engine instance, re #6292
