+++ arches/app/search/search_engine_factory.py
@@ -36,3 +36,5 @@ class SearchEngineFactory(object):
         # _temp = __import__(modulename, globals(), locals(), [classname], -1)
         _temp = __import__(modulename, globals(), locals(), [classname])  # in py3, level must be >= 0
         return getattr(_temp, classname)(hosts=hosts, prefix=prefix, **connection_options)
+
+SearchEngineInstance = SearchEngineFactory().create()
